+++ lib/con_stdio.c
@@ -1,31 +1,21 @@
-
 /* plugins for con_putstr() in conjunction with stdio_portable.c . */
-
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
-
 #include <sneks/console.h>
 
-
-long sneks_con_write(void *cookie, const char *buf, size_t size)
+ssize_t sneks_con_write(void *cookie, const char *buf, size_t size)
 {
-	/* worst possible. revise con_putstr() to something like con_write(),
-	 * where it'll take a size parameter.
-	 */
-	char copy[size + 1];
+	char copy[size + 1]; /* TODO: fugly. replace w/ something that takes a length parameter. */
 	memcpy(copy, buf, size);
 	copy[size] = '\0';
 	con_putstr(copy);
-
 	return size;
 }
 
-
 int sneks_setup_console_stdio(void)
 {
-	FILE *out = fopencookie(NULL, "wb",
-		(cookie_io_functions_t){ .write = &sneks_con_write });
+	FILE *out = fopencookie(NULL, "wb", (cookie_io_functions_t){ .write = &sneks_con_write });
 	if(out == NULL) return -ENOMEM;
 
 	if(stdin != NULL) fclose(stdin);
